Hopefully fix 'comparison between signed and unsigned' warning
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@27431 a1c6a512-1295-4272-9138-f99709370657

+++ apps/codecs/libffmpegFLAC/shndec.c
@@ -423,7 +423,7 @@ int shorten_init(ShortenContext* s, uint8_t *buf, int buf_size)
     get_bits(&s->gb, s->bitindex);
 
     /* shorten signature */
-    if (get_bits_long(&s->gb, 32) != bswap_32(ff_get_fourcc("ajkg"))) {
+    if ((unsigned)get_bits_long(&s->gb, 32) != bswap_32(ff_get_fourcc("ajkg"))) {
         return -1;
     }
